requirements/requirements_tox.txt
black
click
flake8
lxml
mypy
pytest

pyyaml
